feat: add railway stations
and adjust some styles
