+++ src/components/LooMap/LooMap.tsx
@@ -18,19 +18,8 @@ import React from 'react';
 import router from 'next/router';
 import ZoomControl from './ZoomControl';
 import crosshairSvg from '../../../public/crosshair.svg';
-import {
-  CenteredTextSymbolizer,
-  CircleSymbolizer,
-  GroupSymbolizer,
-  Justify,
-  LineSymbolizer,
-  OffsetTextSymbolizer,
-  TextPlacements,
-  labelRules,
-  leafletLayer,
-  paintRules,
-} from 'protomaps-leaflet';
-import { getNumber, getString, mapTheme } from './mapTheme';
+import { leafletLayer } from 'protomaps-leaflet';
+import { labelRules, paintRules } from './mapTheme';
 
 const MapTracker = () => {
   const [, setMapState] = useMapState();
@@ -118,234 +107,8 @@ const LooMap: React.FC<LooMapProps> = ({
     const layer = leafletLayer({
       // Free for non-commercial use https://protomaps.com/
       url: 'https://api.protomaps.com/tiles/v3/{z}/{x}/{y}.mvt?key=73e8a482f059f3f5',
-      paintRules: paintRules(mapTheme).concat({
-        dataLayer: 'roads',
-        symbolizer: new LineSymbolizer({
-          color: '#bbc',
-        }),
-        filter: (z, f) => {
-          const kind = getString(f.props, 'pmap:kind');
-          return kind === 'park';
-        },
-      }),
-      labelRules: labelRules(mapTheme)
-        .slice(0, -2) // remove the default “places_locality” rule
-        .concat(
-          {
-            dataLayer: 'places',
-            minzoom: 9,
-            symbolizer: new CenteredTextSymbolizer({
-              labelProps: ['name'],
-              fill: mapTheme.city_label,
-              stroke: mapTheme.city_label_halo,
-              width: 1,
-              textTransform: (z) => (z > 12 ? 'uppercase' : undefined),
-              lineHeight: 1.5,
-              font: (z, f) => {
-                if (!f) return '400 12px sans-serif';
-                const minZoom = getNumber(f.props, 'pmap:min_zoom');
-                let weight = 400;
-                if (minZoom && minZoom <= 5) {
-                  weight = 600;
-                }
-                let size = 12;
-                const popRank = getNumber(f.props, 'pmap:population_rank');
-                if (popRank && popRank > 9) {
-                  size = 16;
-                }
-                return `${weight} ${size}px sans-serif`;
-              },
-            }),
-            sort: (a, b) => {
-              const aRank = getNumber(a, 'pmap:population_rank');
-              const bRank = getNumber(b, 'pmap:population_rank');
-              return aRank - bRank;
-            },
-            filter: (z, { props }) => {
-              return props['pmap:kind'] === 'locality';
-            },
-          },
-          {
-            dataLayer: 'pois',
-            symbolizer: new GroupSymbolizer([
-              new CircleSymbolizer({
-                fill: '#0a165e',
-                stroke: '#f4f4f4',
-              }),
-              new OffsetTextSymbolizer({
-                labelProps: ['name'],
-                fill: '#889',
-                stroke: '#f4f4f4',
-                width: 1,
-                lineHeight: 0.875,
-                font: '400 16px',
-                offsetY: 4,
-                placements: [TextPlacements.S],
-                justify: Justify.Center,
-              }),
-            ]),
-            filter: (_, { props }) =>
-              [
-                'attraction',
-                'landmark',
-                'memorial',
-                'social_facility',
-                'supermarket',
-              ].includes(getString(props, 'pmap:kind')),
-          },
-          {
-            dataLayer: 'pois',
-            symbolizer: new CenteredTextSymbolizer({
-              labelProps: ['name'],
-              fill: '#0a165e',
-              stroke: 'transparent',
-              width: 1,
-              lineHeight: 1,
-              font: '400 14px',
-            }),
-            filter: (z, { props }) => {
-              if (z < 18) return false;
-              return [
-                'aerodrome',
-                'adult_gaming_centre',
-                'airfield',
-                'alpine_hut',
-                'amusement_ride',
-                'animal',
-                'art',
-                'artwork',
-                'atm',
-                'attraction',
-                'atv',
-                'baby_hatch',
-                'bakery',
-                'bbq',
-                'beauty',
-                'bed_and_breakfast',
-                'bench',
-                'bicycle_parking',
-                'bicycle_rental',
-                'bicycle_repair_station',
-                'boat_storage',
-                'bookmaker',
-                'books',
-                'bureau_de_change',
-                'bus_stop',
-                'butcher',
-                'cafe',
-                'camp_site',
-                'car_parts',
-                'car_rental',
-                'car_repair',
-                'car_sharing',
-                'car_wash',
-                'car',
-                'carousel',
-                'cemetery',
-                'chalet',
-                'charging_station',
-                'childcare',
-                'clinic',
-                'clothes',
-                'college',
-                'computer',
-                'convenience',
-                'customs',
-                'dentist',
-                'district',
-                'doctors',
-                'dog_park',
-                'drinking_water',
-                'emergency_phone',
-                'fashion',
-                'firepit',
-                'fishing',
-                'florist',
-                'forest',
-                'fuel',
-                'gambling',
-                'garden_centre',
-                'gift',
-                'golf_course',
-                'golf',
-                'greengrocer',
-                'grocery',
-                'guest_house',
-                'hairdresser',
-                'hanami',
-                'harbourmaster',
-                'hifi',
-                'hospital',
-                'hostel',
-                'hotel',
-                'hunting_stand',
-                'information',
-                'jewelry',
-                'karaoke_box',
-                'karaoke',
-                'landmark',
-                'library',
-                'life_ring',
-                'lottery',
-                'marina',
-                'maze',
-                'memorial',
-                'military',
-                'mobile_phone',
-                'money_transfer',
-                'motorcycle_parking',
-                'motorcycle',
-                'national_park',
-                'naval_base',
-                'newsagent',
-                'optician',
-                'park',
-                'parking',
-                'perfumery',
-                'picnic_site',
-                'picnic_table',
-                'pitch',
-                'playground',
-                'post_box',
-                'post_office',
-                'ranger_station',
-                'recycling',
-                'roller_coaster',
-                'sanitary_dump_station',
-                'school',
-                'scuba_diving',
-                'shelter',
-                'ship_chandler',
-                'shower',
-                'slipway',
-                'snowmobile',
-                'social_facility',
-                'stadium',
-                'stationery',
-                'studio',
-                'summer_toboggan',
-                'supermarket',
-                'swimming_area',
-                'taxi',
-                'telephone',
-                'tobacco',
-                'toilets',
-                'townhall',
-                'trail_riding_station',
-                'travel_agency',
-                'university',
-                'viewpoint',
-                'waste_basket',
-                'waste_disposal',
-                'water_point',
-                'water_slide',
-                'watering_place',
-                'wayside_cross',
-                'wilderness_hut',
-              ].includes(getString(props, 'pmap:kind'));
-            },
-          },
-        ),
+      paintRules,
+      labelRules,
     });
 
     // @ts-expect-error -- this is what the docs recommend
